extracted file upload component
